Move the "scripts" namespace to its own module.
